Extract common validator logic to the base class
