+++ lib/langchain.rb
@@ -102,6 +102,7 @@ module Langchain
 
   module Utils
     module TokenLength
+      autoload :BaseValidator, "langchain/utils/token_length/base_validator"
       autoload :TokenLimitExceeded, "langchain/utils/token_length/token_limit_exceeded"
       autoload :OpenAIValidator, "langchain/utils/token_length/openai_validator"
       autoload :GooglePalmValidator, "langchain/utils/token_length/google_palm_validator"
