+++ lib/langchain/utils/token_length/openai_validator.rb
@@ -9,7 +9,7 @@ module Langchain
       # This class is meant to validate the length of the text passed in to OpenAI's API.
       # It is used to validate the token length before the API call is made
       #
-      class OpenAIValidator
+      class OpenAIValidator < BaseValidator
         TOKEN_LIMITS = {
           # Source:
           # https://platform.openai.com/docs/api-reference/embeddings
@@ -38,44 +38,19 @@ module Langchain
         }.freeze
 
         #
-        # Calculate the `max_tokens:` parameter to be set by calculating the context length of the text minus the prompt length
-        #
-        # @param content [String | Array<String>] The text or array of texts to validate
-        # @param model_name [String] The model name to validate against
-        # @return [Integer] Whether the text is valid or not
-        # @raise [TokenLimitExceeded] If the text is too long
-        #
-        def self.validate_max_tokens!(content, model_name)
-          text_token_length = if content.is_a?(Array)
-            content.sum { |item| token_length(item.to_json, model_name) }
-          else
-            token_length(content, model_name)
-          end
-
-          max_tokens = TOKEN_LIMITS[model_name] - text_token_length
-
-          # Raise an error even if whole prompt is equal to the model's token limit (max_tokens == 0) since not response will be returned
-          if max_tokens <= 0
-            raise limit_exceeded_exception(TOKEN_LIMITS[model_name], text_token_length)
-          end
-
-          max_tokens
-        end
-
-        #
         # Calculate token length for a given text and model name
         #
         # @param text [String] The text to calculate the token length for
         # @param model_name [String] The model name to validate against
         # @return [Integer] The token length of the text
         #
-        def self.token_length(text, model_name)
+        def self.token_length(text, model_name, options = {})
           encoder = Tiktoken.encoding_for_model(model_name)
           encoder.encode(text).length
         end
 
-        def self.limit_exceeded_exception(limit, length)
-          TokenLimitExceeded.new("This model's maximum context length is #{limit} tokens, but the given text is #{length} tokens long.", length - limit)
+        def self.token_limit(model_name)
+          TOKEN_LIMITS[model_name]
         end
       end
     end
